Bug Fix
-> Put different functionalities in different files.
-> Fix a bug on ArrowDown, "commandIndx" in if condition.
