+++ apps/macos-portfolio/package.json
@@ -1,6 +1,6 @@
 {
   "name": "macos-portfolio",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "private": true,
   "scripts": {
     "dev": "next dev",
